change- Moved IGet/IPost to api and swapped project references
