+++ api/api.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\HTTPServer\HTTPServer.csproj" />
-  </ItemGroup>
-
 </Project>
